+++ dbmodel/utils/fct/gw_fct_getwidgetvalues.sql
@@ -57,23 +57,23 @@ BEGIN
 	v_tgop = ((p_data ->>'data')::json->>'tgOp');
 
 	IF v_featuretype ='NODE' THEN
-	
-		v_min = (SELECT max(y) FROM (SELECT sys_y1 as y FROM v_edit_arc WHERE node_1=v_id UNION SELECT sys_y2 FROM v_edit_arc WHERE node_2=v_id)a);
+
+		v_min = (SELECT max(y) FROM (SELECT sys_y1 as y FROM v_edit_arc WHERE node_1=v_id::integer UNION SELECT sys_y2 FROM v_edit_arc WHERE node_2=v_id::integer)a);
 		IF v_min IS NULL THEN v_min = 0; END IF;
-		
+
 	ELSIF v_featuretype ='ARC' THEN
-			
-		SELECT * INTO v_noderecord1 FROM v_edit_node WHERE node_id=v_node_1;
+
+		SELECT * INTO v_noderecord1 FROM v_edit_node WHERE node_id=v_node_1::integer;
 		v_max1 = v_noderecord1.sys_ymax;
 		IF v_max1 IS NULL THEN v_max1 = 999; END IF;
 
-		
-		SELECT * INTO v_noderecord2 FROM v_edit_node WHERE node_id=v_node_2;
+
+		SELECT * INTO v_noderecord2 FROM v_edit_node WHERE node_id=v_node_2::integer;
 		v_max2 = v_noderecord2.sys_ymax;
 		IF v_max2 IS NULL THEN v_max2 = 999; END IF;
-					
+
 	END IF;
-		         
+
 	-- Return
 	RETURN ('{"ymax":{"min":'||v_min||',"max":'||v_max||'},'||
 	  '"custom_ymax":{"min":'||v_min||',"max":'||v_max||'},'||
@@ -81,7 +81,7 @@ BEGIN
 		'"custom_y1":{"min":'||v_min1||',"max":'||v_max1||'},'||
 			   '"y2":{"min":'||v_min2||',"max":'||v_max2||'},'||
 		'"custom_y2":{"min":'||v_min2||',"max":'||v_max2||'}}');
-   
+
 END;
 $BODY$
   LANGUAGE plpgsql VOLATILE
